Remove deprecated API reference files and logging configuration. This commit deletes the following files: api_refs.py, fetch_points.py, logger_config.py, main.py, parser.py, and print_data.py, which were previously used for API interactions, data fetching, and logging setup. The removal is part of a refactor to streamline the codebase and improve maintainability.
